Rename pwb debug to configuration.
