+++ packages/kartoffelgames.web.potato_web_builder/test/integration/module/event-attribute.module.test.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import { PwbComponent } from '../../../source/core/component/pwb-component.decorator';
-import { PwbConfiguration } from '../../../source/core/configuration/pwb-debug';
+import { PwbConfiguration } from '../../../source/core/configuration/pwb-configuration';
 import { Processor } from '../../../source/core/core_entity/processor';
 import '../../mock/request-animation-frame-mock-session';
 import '../../utility/chai-helper';
