+++ packages/kartoffelgames.web.potato_web_builder/test/integration/module/slot-instruction-module.test.ts
@@ -1,7 +1,7 @@
 import { InjectionConstructor } from '@kartoffelgames/core.dependency-injection';
 import { expect } from 'chai';
 import { PwbComponent } from '../../../source/core/component/pwb-component.decorator';
-import { PwbConfiguration } from '../../../source/core/configuration/pwb-debug';
+import { PwbConfiguration } from '../../../source/core/configuration/pwb-configuration';
 import { Processor } from '../../../source/core/core_entity/processor';
 import '../../mock/request-animation-frame-mock-session';
 import '../../utility/chai-helper';
